Used MVP Feedback to improve codebase:
